big ol refactor
